+++ packages/icon-view/src/components/index.ts
@@ -1,3 +1,5 @@
 export * from './super-ellipse';
 export * from './circle';
+export * from './rectangle';
+export * from './no-shape';
 export * from './base-shape';