+++ packages/input-autocomplete/src/mobile.ts
@@ -1,1 +1,2 @@
 export * from './Component.mobile';
+export * from './Component.modal.mobile';