+++ packages/modal/src/components/footer/Component.tsx
@@ -59,6 +59,7 @@ export const Footer: FC<FooterProps> = ({ children, className, sticky, layout = 
                     [desktopStyles[size]]: view === 'desktop',
                     [mobileStyles.footer]: view === 'mobile',
                     [mobileStyles.sticky]: view === 'mobile' && sticky,
+                    [layoutStyles[`${layout}-mobile`]]: view === 'mobile',
                 },
             )}
         >