+++ packages/select/src/components/base-select-mobile/Component.tsx
@@ -59,6 +59,21 @@ export type SelectMobileProps = Omit<BaseSelectProps, 'Checkmark' | 'onScroll'> 
      * Дополнительные пропсы шторки
      */
     bottomSheetProps?: Partial<BottomSheetProps>;
+
+    /**
+     *  Дополнительные пропсы модалки
+     */
+    modalProps?: Partial<React.ComponentProps<typeof ModalMobile>>;
+
+    /**
+     *  Дополнительные пропсы шапки модалки
+     */
+    modalHeaderProps?: Partial<React.ComponentProps<typeof ModalMobile.Header>>;
+
+    /**
+     *  Дополнительные пропсы футера модалки
+     */
+    modalFooterProps?: Partial<React.ComponentProps<typeof ModalMobile.Footer>>;
 };
 
 export const BaseSelectMobile = forwardRef(
@@ -107,6 +122,9 @@ export const BaseSelectMobile = forwardRef(
             footer,
             isBottomSheet,
             bottomSheetProps,
+            modalProps,
+            modalHeaderProps,
+            modalFooterProps,
             showEmptyOptionsList = false,
         }: SelectMobileProps,
         ref,
@@ -115,6 +133,7 @@ export const BaseSelectMobile = forwardRef(
         const fieldRef = useRef<HTMLInputElement>(null);
         const listRef = useRef<HTMLDivElement>(null);
         const initiatorRef = useRef<OptionShape | null>(null);
+        const alreadyClickedRef = useRef<boolean>(false);
 
         const itemToString = (option: OptionShape) => (option ? option.key : '');
 
@@ -211,7 +230,12 @@ export const BaseSelectMobile = forwardRef(
                     case useCombobox.stateChangeTypes.ItemClick:
                         initiatorRef.current = selectedItem;
 
-                        if (selectedItem && !selectedItem.disabled) {
+                        if (selectedItem && !selectedItem.disabled && !alreadyClickedRef.current) {
+                            alreadyClickedRef.current = true;
+                            // TODO!!! Проблема downshift + React 18. ItemClick срабатывает дважды. См https://github.com/downshift-js/downshift/issues/1384
+                            setTimeout(() => {
+                                alreadyClickedRef.current = false;
+                            });
                             const alreadySelected = selectedItems.includes(selectedItem);
                             const allowRemove =
                                 allowUnselect || (multiple && selectedItems.length > 1);
@@ -427,16 +451,25 @@ export const BaseSelectMobile = forwardRef(
                 ) : (
                     <ModalMobile
                         open={open}
-                        onClose={handleClose}
-                        contentClassName={styles.sheetContent}
                         hasCloser={true}
-                        ref={scrollableContainerRef}
+                        {...modalProps}
+                        onClose={(...args) => {
+                            handleClose();
+                            modalProps?.onClose?.(...args);
+                        }}
+                        contentClassName={cn(styles.sheetContent, modalProps?.contentClassName)}
+                        ref={mergeRefs([
+                            scrollableContainerRef,
+                            modalProps?.ref as React.Ref<HTMLDivElement>,
+                        ])}
                     >
-                        <ModalMobile.Header hasCloser={true} sticky={true}>
+                        <ModalMobile.Header hasCloser={true} sticky={true} {...modalHeaderProps}>
                             {label || placeholder}
                         </ModalMobile.Header>
 
                         {renderOptionsList()}
+
+                        {modalFooterProps?.children && <ModalMobile.Footer {...modalFooterProps} />}
                     </ModalMobile>
                 )}
             </div>