+++ packages/side-panel/src/components/footer/Component.mobile.tsx
@@ -7,12 +7,18 @@ import styles from './mobile.module.css';
 
 export type FooterMobileProps = FooterProps;
 
-export const FooterMobile: FC<FooterMobileProps> = ({ className, sticky, ...restProps }) => (
+export const FooterMobile: FC<FooterMobileProps> = ({
+    className,
+    sticky,
+    layout = 'start',
+    ...restProps
+}) => (
     <Footer
-        className={cn(className, styles.footer, {
+        className={cn(className, styles.footer, styles[layout], {
             [styles.sticky]: sticky,
         })}
         sticky={sticky}
+        layout={layout}
         {...restProps}
     />
 );