+++ packages/keyboard-focusable/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.0.6
+
+### Patch Changes
+
+### [#676](https://github.com/core-ds/core-components/pull/676)
+
+-   Обновлена зависимость @alfalab/hooks
+
 ## 4.0.5
 
 ### Patch Changes