+++ packages/tag/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 5.3.2
+
+### Patch Changes
+
+### [#676](https://github.com/core-ds/core-components/pull/676)
+
+-   Обновлена зависимость @alfalab/hooks
+
 ## 5.3.1
 
 ### Patch Changes