+++ packages/time-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alfalab/core-components-time-input
 
+## 2.1.17
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@11.1.16
+
 ## 2.1.16
 
 ### Patch Changes